Split the program across multiple files
