+++ src/error_handling.c
@@ -1,0 +1,10 @@
+#include "headers/includes.h"
+
+void die(const char *s)
+{
+  write(STDOUT_FILENO, "\x1b[2J", 4);
+  write(STDOUT_FILENO, "\x1b[H", 3);
+
+  perror(s);
+  exit(1);
+}
