+++ src/headers/includes.h
@@ -1,0 +1,19 @@
+#ifndef INCLUDES
+#define INCLUDES
+
+#include <stdio.h>
+#include <ctype.h>
+#include <errno.h>
+#include <stdlib.h>
+#include <string.h>
+#include <stdarg.h>
+#include <sys/ioctl.h>
+#include <sys/types.h>
+#include <unistd.h>
+#include <termios.h>
+#include <time.h>
+#include <assert.h>
+#include <fcntl.h>
+#include <stdbool.h>
+
+#endif
